Complete dual-target RISC-V regression

+++ examples/kernels/quantization/mul_mat.py
@@ -202,7 +202,7 @@ def production_mul_mat_q6_k(
 @weft.kernel
 def production_mul_mat_iq1_s(
     W: View[IQ1_S, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    grid: View[f32, (16384,)], Y: View[f32, (M, N)]
+    grid: View[i8, (16384,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq1_s(W, X, Xq, grid, Y)
 
@@ -210,7 +210,7 @@ def production_mul_mat_iq1_s(
 @weft.kernel
 def production_mul_mat_iq1_m(
     W: View[IQ1_M, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    grid: View[f32, (16384,)], f16_bits: View[f32, (65536,)], Y: View[f32, (M, N)]
+    grid: View[i8, (16384,)], f16_bits: View[f32, (65536,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq1_m(W, X, Xq, grid, f16_bits, Y)
 
@@ -218,7 +218,7 @@ def production_mul_mat_iq1_m(
 @weft.kernel
 def production_mul_mat_iq2_s(
     W: View[IQ2_S, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    grid: View[f32, (8192,)], Y: View[f32, (M, N)]
+    grid: View[i8, (8192,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq2_s(W, X, Xq, grid, Y)
 
@@ -226,7 +226,7 @@ def production_mul_mat_iq2_s(
 @weft.kernel
 def production_mul_mat_iq2_xs(
     W: View[IQ2_XS, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    grid: View[f32, (4096,)], signs: View[f32, (1024,)], Y: View[f32, (M, N)]
+    grid: View[i8, (4096,)], signs: View[i8, (1024,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq2_xs(W, X, Xq, grid, signs, Y)
 
@@ -234,7 +234,7 @@ def production_mul_mat_iq2_xs(
 @weft.kernel
 def production_mul_mat_iq2_xxs(
     W: View[IQ2_XXS, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    grid: View[f32, (2048,)], signs: View[f32, (1024,)], Y: View[f32, (M, N)]
+    grid: View[i8, (2048,)], signs: View[i8, (1024,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq2_xxs(W, X, Xq, grid, signs, Y)
 
@@ -254,7 +254,7 @@ def production_mul_mat_iq2_xxs_staged(
 @weft.kernel
 def production_mul_mat_iq3_s(
     W: View[IQ3_S, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    grid: View[f32, (2048,)], Y: View[f32, (M, N)]
+    grid: View[i8, (2048,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq3_s(W, X, Xq, grid, Y)
 
@@ -262,7 +262,7 @@ def production_mul_mat_iq3_s(
 @weft.kernel
 def production_mul_mat_iq3_xxs(
     W: View[IQ3_XXS, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    grid: View[f32, (1024,)], signs: View[f32, (1024,)], Y: View[f32, (M, N)]
+    grid: View[i8, (1024,)], signs: View[i8, (1024,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq3_xxs(W, X, Xq, grid, signs, Y)
 
@@ -270,7 +270,7 @@ def production_mul_mat_iq3_xxs(
 @weft.kernel
 def production_mul_mat_iq4_nl(
     W: View[IQ4_NL, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_0, (M, K)],
-    codebook: View[f32, (16,)], Y: View[f32, (M, N)]
+    codebook: View[i8, (16,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq4_nl(W, X, Xq, codebook, Y)
 
@@ -278,7 +278,7 @@ def production_mul_mat_iq4_nl(
 @weft.kernel
 def production_mul_mat_iq4_xs(
     W: View[IQ4_XS, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_K, (M, K)],
-    codebook: View[f32, (16,)], Y: View[f32, (M, N)]
+    codebook: View[i8, (16,)], Y: View[f32, (M, N)]
 ):
     mul_mat_iq4_xs(W, X, Xq, codebook, Y)
 
@@ -301,7 +301,7 @@ def production_mul_mat_tq2_0(
 @weft.kernel
 def production_mul_mat_mxfp4(
     W: View[MXFP4, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_0, (M, K)],
-    codebook: View[f32, (16,)], scale: View[f32, (256,)], Y: View[f32, (M, N)]
+    codebook: View[i8, (16,)], scale: View[f32, (256,)], Y: View[f32, (M, N)]
 ):
     mul_mat_mxfp4(W, X, Xq, codebook, scale, Y)
 
@@ -309,6 +309,6 @@ def production_mul_mat_mxfp4(
 @weft.kernel
 def production_mul_mat_nvfp4(
     W: View[NVFP4, (N, K)], X: View[f32, (M, K)], Xq: View[Q8_0, (M, K)],
-    codebook: View[f32, (16,)], scale: View[f32, (256,)], Y: View[f32, (M, N)]
+    codebook: View[i8, (16,)], scale: View[f32, (256,)], Y: View[f32, (M, N)]
 ):
     mul_mat_nvfp4(W, X, Xq, codebook, scale, Y)
