+++ examples/kernels/quantization/row_dequantize.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import weft
-from weft.language import View, f32, u32
+from weft.language import View, f32, i8, u32
 from weft.std.encodings import (
     IQ1_M,
     IQ1_S,
@@ -113,7 +113,7 @@ def row_dequantize_q6_k(W: View[Q6_K, (K,)], Y: View[f32, (K,)]):
 
 @weft.kernel
 def row_dequantize_iq1_s(
-    W: View[IQ1_S, (K,)], grid: View[f32, (16384,)], Y: View[f32, (K,)]
+    W: View[IQ1_S, (K,)], grid: View[i8, (16384,)], Y: View[f32, (K,)]
 ):
     dequantize_iq1_s(W, grid, Y)
 
@@ -121,7 +121,7 @@ def row_dequantize_iq1_s(
 @weft.kernel
 def row_dequantize_iq1_m(
     W: View[IQ1_M, (K,)],
-    grid: View[f32, (16384,)],
+    grid: View[i8, (16384,)],
     f16_bits: View[f32, (65536,)],
     Y: View[f32, (K,)],
 ):
@@ -130,7 +130,7 @@ def row_dequantize_iq1_m(
 
 @weft.kernel
 def row_dequantize_iq2_s(
-    W: View[IQ2_S, (K,)], grid: View[f32, (8192,)], Y: View[f32, (K,)]
+    W: View[IQ2_S, (K,)], grid: View[i8, (8192,)], Y: View[f32, (K,)]
 ):
     dequantize_iq2_s(W, grid, Y)
 
@@ -138,8 +138,8 @@ def row_dequantize_iq2_s(
 @weft.kernel
 def row_dequantize_iq2_xs(
     W: View[IQ2_XS, (K,)],
-    grid: View[f32, (4096,)],
-    signs: View[f32, (1024,)],
+    grid: View[i8, (4096,)],
+    signs: View[i8, (1024,)],
     Y: View[f32, (K,)],
 ):
     dequantize_iq2_xs(W, grid, signs, Y)
@@ -148,8 +148,8 @@ def row_dequantize_iq2_xs(
 @weft.kernel
 def row_dequantize_iq2_xxs(
     W: View[IQ2_XXS, (K,)],
-    grid: View[f32, (2048,)],
-    signs: View[f32, (1024,)],
+    grid: View[i8, (2048,)],
+    signs: View[i8, (1024,)],
     Y: View[f32, (K,)],
 ):
     dequantize_iq2_xxs(W, grid, signs, Y)
@@ -157,7 +157,7 @@ def row_dequantize_iq2_xxs(
 
 @weft.kernel
 def row_dequantize_iq3_s(
-    W: View[IQ3_S, (K,)], grid: View[f32, (2048,)], Y: View[f32, (K,)]
+    W: View[IQ3_S, (K,)], grid: View[i8, (2048,)], Y: View[f32, (K,)]
 ):
     dequantize_iq3_s(W, grid, Y)
 
@@ -165,8 +165,8 @@ def row_dequantize_iq3_s(
 @weft.kernel
 def row_dequantize_iq3_xxs(
     W: View[IQ3_XXS, (K,)],
-    grid: View[f32, (1024,)],
-    signs: View[f32, (1024,)],
+    grid: View[i8, (1024,)],
+    signs: View[i8, (1024,)],
     Y: View[f32, (K,)],
 ):
     dequantize_iq3_xxs(W, grid, signs, Y)
@@ -174,14 +174,14 @@ def row_dequantize_iq3_xxs(
 
 @weft.kernel
 def row_dequantize_iq4_nl(
-    W: View[IQ4_NL, (K,)], codebook: View[f32, (16,)], Y: View[f32, (K,)]
+    W: View[IQ4_NL, (K,)], codebook: View[i8, (16,)], Y: View[f32, (K,)]
 ):
     dequantize_iq4_nl(W, codebook, Y)
 
 
 @weft.kernel
 def row_dequantize_iq4_xs(
-    W: View[IQ4_XS, (K,)], codebook: View[f32, (16,)], Y: View[f32, (K,)]
+    W: View[IQ4_XS, (K,)], codebook: View[i8, (16,)], Y: View[f32, (K,)]
 ):
     dequantize_iq4_xs(W, codebook, Y)
 
@@ -201,7 +201,7 @@ def row_dequantize_tq2_0(W: View[TQ2_0, (K,)], Y: View[f32, (K,)]):
 @weft.kernel
 def row_dequantize_mxfp4(
     W: View[MXFP4, (K,)],
-    codebook: View[f32, (16,)],
+    codebook: View[i8, (16,)],
     scale: View[f32, (256,)],
     Y: View[f32, (K,)],
 ):
@@ -211,7 +211,7 @@ def row_dequantize_mxfp4(
 @weft.kernel
 def row_dequantize_nvfp4(
     W: View[NVFP4, (K,)],
-    codebook: View[f32, (16,)],
+    codebook: View[i8, (16,)],
     scale: View[f32, (256,)],
     Y: View[f32, (K,)],
 ):
